Migrate PowerOutputGroup to j5
