+++ j5/boards/sr/v4/power_board.py
@@ -1,11 +1,18 @@
 """Classes for the SR v4 Power Board."""
 
 from enum import Enum
-from typing import TYPE_CHECKING, List, Union, cast, Mapping
+from typing import TYPE_CHECKING, List, Mapping, cast
 
 from j5.backends import Backend, Environment
 from j5.boards import Board
-from j5.components import LED, BatterySensor, Button, Piezo, PowerOutput
+from j5.components import (
+    LED,
+    BatterySensor,
+    Button,
+    Piezo,
+    PowerOutput,
+    PowerOutputGroup,
+)
 
 if TYPE_CHECKING:
     from j5.components import (  # noqa: F401
@@ -34,38 +41,6 @@ class PowerOutputPosition(Enum):
     L3 = 5
 
 
-class PowerOutputGroup:
-    """A group of PowerOutputs on the PowerBoard."""
-
-    def __init__(self, backend: Backend, board: Board):
-        self._backend = backend
-        self._board = board
-
-        self._outputs: Mapping[PowerOutputPosition, PowerOutput] = {
-            output: PowerOutput(output.value, self._board, cast("PowerOutputInterface", self._backend))
-            for output in PowerOutputPosition
-            # Note that in Python 3, Enums are ordered.
-        }
-
-    def power_on(self) -> None:
-        """Enable all outputs in the group."""
-        for output in self._outputs:
-            output.is_enabled = True
-
-    def power_off(self) -> None:
-        """Disable all outputs in the group."""
-        for output in self._outputs:
-            output.is_enabled = False
-
-    def __getitem__(self, index: PowerOutputPosition) -> PowerOutput:
-        """Get an output using list notation."""
-        return self._outputs[index]
-
-    def __len__(self) -> int:
-        """Get the number of outputs in the group."""
-        return len(self._outputs)
-
-
 class PowerBoard(Board):
     """Student Robotics v4 Power Board."""
 
@@ -75,7 +50,16 @@ class PowerBoard(Board):
 
         self._backend = self._environment.get_backend(self.__class__)
 
-        self._output_group = PowerOutputGroup(self._backend, self)
+        self._outputs: Mapping[PowerOutputPosition, PowerOutput] = {
+            output: PowerOutput(
+                output.value, self, cast("PowerOutputInterface", self._backend),
+            )
+            for output in PowerOutputPosition
+            # Note that in Python 3, Enums are ordered.
+        }
+
+        self._output_group = PowerOutputGroup(self._outputs)
+
         self._piezo = Piezo(0, self, cast("PiezoInterface", self._backend))
         self._start_button = Button(0, self, cast("ButtonInterface", self._backend))
         self._battery_sensor = BatterySensor(
